+++ packages/gatsby/cache-dir/api-ssr-docs.js
@@ -99,17 +99,17 @@ exports.onRenderBody = true
  * replace head components to be rendered in your `html.js`. This is useful if
  * you need to reorder scripts or styles added by other plugins.
  * @param {Object} $0
- * @param {Array} $0.getHeadComponents Returns the current `headComponents` array.
+ * @param {Array<ReactNode>} $0.getHeadComponents Returns the current `headComponents` array.
  * @param {function} $0.replaceHeadComponents Takes an array of components as its
  * first argument which replace the `headComponents` array which is passed
  * to the `html.js` component. **WARNING** if multiple plugins implement this
  * API it's the last plugin that "wins".
- * @param {Array} $0.getPreBodyComponents Returns the current `preBodyComponents` array.
+ * @param {Array<ReactNode>} $0.getPreBodyComponents Returns the current `preBodyComponents` array.
  *  @param {function} $0.replacePreBodyComponents Takes an array of components as its
  * first argument which replace the `preBodyComponents` array which is passed
  * to the `html.js` component. **WARNING** if multiple plugins implement this
  * API it's the last plugin that "wins".
- * @param {Array} $0.getPostBodyComponents Returns the current `postBodyComponents` array.
+ * @param {Array<ReactNode>} $0.getPostBodyComponents Returns the current `postBodyComponents` array.
  *  @param {function} $0.replacePostBodyComponents Takes an array of components as its
  * first argument which replace the `postBodyComponents` array which is passed
  * to the `html.js` component. **WARNING** if multiple plugins implement this
@@ -140,9 +140,8 @@ exports.onPreRenderHTML = true
  *
  * _Note:_ [There is equivalent hook in Browser API](/docs/browser-apis/#wrapPageElement)
  * @param {object} $0
- * @param {object} $0.element The "Page" React Element built by Gatsby.
+ * @param {ReactNode} $0.element The "Page" React Element built by Gatsby.
  * @param {object} $0.props Props object used by page.
- * @param {string} $0.pathname Path of page.
  * @example
  * import React from "react"
  * import Layout from "./src/components/layout"
@@ -163,7 +162,7 @@ exports.wrapPageElement = true
  *
  * _Note:_ [There is equivalent hook in Browser API](/docs/browser-apis/#wrapRootElement)
  * @param {object} $0
- * @param {object} $0.element The "Root" React Element built by Gatsby.
+ * @param {ReactNode} $0.element The "Root" React Element built by Gatsby.
  * @example
  * import React from "react"
  * import { Provider } from "react-redux"