+++ packages/gatsby/index.d.ts
@@ -1,4 +1,5 @@
 import * as React from "react"
+import { WindowLocation } from "@reach/router"
 
 export {
   default as Link,
@@ -23,3 +24,11 @@ export class StaticQuery extends React.Component<StaticQueryProps> {}
 export const useStaticQuery: <TData = any>(query: any) => TData
 
 export const graphql: (query: TemplateStringsArray) => void
+
+export const parsePath: (path: string) => WindowLocation
+
+export interface PageRendererProps {
+  location: WindowLocation
+}
+
+export class PageRenderer extends React.Component<PageRendererProps> {}