+++ packages/gatsby/src/internal-plugins/query-runner/query-compiler.js
@@ -12,6 +12,7 @@ import filterContextForNode from "@gatsbyjs/relay-compiler/lib/filterContextForN
 const _ = require(`lodash`)
 
 import { store } from "../../redux"
+const { boundActionCreators } = require(`../../redux/actions`)
 import FileParser from "./file-parser"
 import GraphQLIRPrinter from "@gatsbyjs/relay-compiler/lib/GraphQLIRPrinter"
 import {
@@ -62,11 +63,9 @@ const validationRules = [
 let lastRunHadErrors = null
 const overlayErrorID = `graphql-compiler`
 
-const resolveThemes = (plugins = []) =>
-  plugins.reduce((merged, plugin) => {
-    if (plugin.name.includes(`gatsby-theme-`)) {
-      merged.push(plugin.resolve)
-    }
+const resolveThemes = (themes = []) =>
+  themes.reduce((merged, theme) => {
+    merged.push(theme.themeDir)
     return merged
   }, [])
 
@@ -136,6 +135,7 @@ class Runner {
     const compiledNodes: Queries = new Map()
     const namePathMap = new Map()
     const nameDefMap = new Map()
+    const nameErrorMap = new Map()
     const documents = []
 
     for (let [filePath, doc] of nodes.entries()) {
@@ -143,6 +143,9 @@ class Runner {
 
       if (errors && errors.length) {
         this.reportError(graphqlValidationError(errors, filePath))
+        boundActionCreators.queryExtractionGraphQLError({
+          componentPath: filePath,
+        })
         return compiledNodes
       }
 
@@ -165,8 +168,18 @@ class Runner {
         )
       )
     } catch (error) {
-      this.reportError(graphqlError(namePathMap, nameDefMap, error))
-      return compiledNodes
+      const { formattedMessage, docName, message, codeBlock } = graphqlError(
+        namePathMap,
+        nameDefMap,
+        error
+      )
+      nameErrorMap.set(docName, { formattedMessage, message, codeBlock })
+      boundActionCreators.queryExtractionGraphQLError({
+        componentPath: namePathMap.get(docName),
+        error: formattedMessage,
+      })
+      this.reportError(formattedMessage)
+      return false
     }
 
     // relay-compiler v1.5.0 added "StripUnusedVariablesTransform" to
@@ -193,6 +206,9 @@ class Runner {
             otherNode && nameDefMap.get(otherNode.name)
           )
         )
+        boundActionCreators.queryExtractionGraphQLError({
+          componentPath: filePath,
+        })
         return
       }
 
@@ -248,9 +264,13 @@ export { Runner, resolveThemes }
 
 export default async function compile(): Promise<Map<string, RootQuery>> {
   // TODO: swap plugins to themes
-  const { program, schema, plugins } = store.getState()
+  const { program, schema, themes } = store.getState()
 
-  const runner = new Runner(program.directory, resolveThemes(plugins), schema)
+  const runner = new Runner(
+    program.directory,
+    resolveThemes(themes.themes),
+    schema
+  )
 
   const queries = await runner.compileAll()