+++ www/src/pages/docs/index.js
@@ -25,61 +25,81 @@ class IndexRoute extends React.Component {
             <p>There are four main ways to get started with Gatsby:</p>
             <ol>
               <li>
-                <Link to="/tutorial/">Tutorial</Link>: The tutorial is written
-                to be as accessible as possible to people without much web
-                development experience.
+                <Link to="/tutorial/">Tutorial</Link>: Step-by-step instructions
+                on how to install Gatsby and start a project: written for people
+                without Gatsby or web development experience, though it has
+                helped developers of all skill levels.
               </li>
               <li>
-                <Link to="/docs/quick-start">Quick start</Link>: The quick start
-                is intended for intermediate to advanced developers who prefer
-                to dig straight in.
+                <Link to="/docs/quick-start">Quick start</Link>: One page
+                summary of how to install Gatsby and start a new project.
               </li>
               <li>
-                <Link to="/docs/recipes">Recipes</Link>: A happy medium between
-                the tutorial and the quick start, find some quick answers for
+                <Link to="/docs/recipes/">Recipes</Link>: A happy medium between
+                the tutorial and the quick start. Find some quick answers for
                 how to accomplish some specific, common tasks with Gatsby.
               </li>
               <li>
                 Choose your own adventure and peruse the various sections of the
                 Gatsby docs:
+                <ul>
+                  <li>
+                    <Link to="/docs/guides/">Guides</Link>: Dive deeper into
+                    different topics around building with Gatsby, like sourcing
+                    data, deployment, and more.
+                  </li>
+                  <li>
+                    <Link to="/ecosystem/">Ecosystem</Link>: Check out libraries
+                    for Gatsby starters and plugins, as well as external
+                    community resources.
+                  </li>
+                  <li>
+                    <Link to="/docs/api-reference/">API Reference</Link>: Learn
+                    more about Gatsby APIs and configuration.
+                  </li>
+                  <li>
+                    <Link to="/docs/releases-and-migration/">
+                      Releases &amp; Migration
+                    </Link>
+                    : Find release notes and guides for migration between major
+                    versions.
+                  </li>
+                  <li>
+                    <Link to="/docs/conceptual-guide/">Conceptual Guide</Link>:
+                    Read high-level overviews of the Gatsby approach.
+                  </li>
+                  <li>
+                    <Link to="/docs/behind-the-scenes/">Behind the Scenes</Link>
+                    : Dig into how Gatsby works under the hood.
+                  </li>
+                  <li>
+                    <Link to="/docs/advanced-tutorials/">
+                      Advanced Tutorials
+                    </Link>
+                    : Learn about topics that are too large for a doc and
+                    warrant a tutorial.
+                  </li>
+                  <li>
+                    <Link to="/docs/using-gatsby-professionally/">
+                      Using Gatsby Professionally
+                    </Link>
+                    : Learn tips and tricks for how to explain Gatsby to others
+                    at work, so that you have more opportunities to work with
+                    Gatsby professionally.
+                  </li>
+                  <li>
+                    <Link to="/contributing/">Contributing</Link>: Find guides
+                    on the Gatsby.js community, code of conduct, and how to get
+                    started contributing.
+                  </li>
+                </ul>
               </li>
-              <ul>
-                <li>
-                  <strong>Guides</strong>: Dive deeper into different topics
-                  around building with Gatsby, like sourcing data, deployment,
-                  and more.
-                </li>
-                <li>
-                  <strong>Ecosystem</strong>: Check out libraries for Gatsby
-                  starters and plugins, as well as external community resources.
-                </li>
-                <li>
-                  <strong>API Reference</strong>: Learn more about Gatsby APIs
-                  and configuration.
-                </li>
-                <li>
-                  <strong>Releases &amp; Migration</strong>: Find release notes
-                  and guides for migration between major versions.
-                </li>
-                <li>
-                  <strong>Conceptual Guide</strong>: Read high level overviews
-                  of the Gatsby approach.
-                </li>
-                <li>
-                  <strong>Behind the Scenes</strong>: Dig into how Gatsby works
-                  under the hood.
-                </li>
-                <li>
-                  <strong>Advanced Tutorials</strong>: Learn about topics that
-                  are too large for a doc and warrant a tutorial.
-                </li>
-                <li>
-                  <strong>Contributing</strong>: Find guides on the Gatsby
-                  community, code of conduct, and how to get started
-                  contributing to Gatsby.
-                </li>
-              </ul>
             </ol>
+            <p>
+              Visit the <Link to="/contributing">Contributing</Link> section to
+              find guides on the Gatsby community, code of conduct, and how to
+              get started contributing to Gatsby.
+            </p>
             <EmailCaptureForm signupMessage="Want to keep up with the latest tips &amp; tricks? Subscribe to our newsletter!" />
           </Container>
         </DocSearchContent>