+++ www/src/views/starter-library/lhs-filter.js
@@ -4,7 +4,6 @@ import React from "react"
 import MdCheckboxBlank from "react-icons/lib/md/check-box-outline-blank"
 import MdCheckbox from "react-icons/lib/md/check-box"
 import { colors } from "../../utils/presets"
-import { scale } from "../../utils/typography"
 
 import Collapsible from "../shared/collapsible"
 import styles from "../shared/styles"
@@ -46,8 +45,7 @@ export default function LHSFilter({
               <div
                 css={{
                   color: filters.has(c) ? colors.gatsby : colors.ui.bright,
-                  ...scale(0),
-                  marginRight: 8,
+                  ...styles.filterCheckbox,
                 }}
               >
                 {filters.has(c) ? (