+++ packages/babel-plugin-remove-graphql-queries/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.2](https://github.com/gatsbyjs/gatsby/compare/babel-plugin-remove-graphql-queries@2.6.1...babel-plugin-remove-graphql-queries@2.6.2) (2019-03-11)
+
+**Note:** Version bump only for package babel-plugin-remove-graphql-queries
+
+## [2.6.1](https://github.com/gatsbyjs/gatsby/compare/babel-plugin-remove-graphql-queries@2.6.0...babel-plugin-remove-graphql-queries@2.6.1) (2019-02-22)
+
+**Note:** Version bump only for package babel-plugin-remove-graphql-queries
+
 # [2.6.0](https://github.com/gatsbyjs/gatsby/compare/babel-plugin-remove-graphql-queries@2.5.5...babel-plugin-remove-graphql-queries@2.6.0) (2019-02-13)
 
 ### Features